refactor: GenreWeightsをmodels/からconstants/に移動
- dataclassではなく定数クラスであるため、constantsパッケージがより適切
- src/constants/ディレクトリを新規作成
- すべてのインポートパスを更新
