+++ src/constants/__init__.py
@@ -1,0 +1,1 @@
+"""Constants package for genre weights and other constants."""
