+++ tests/analyzers/test_metric_calculator.py
@@ -21,8 +21,8 @@ from PIL import Image
 
 from src.analyzers.clip_model_manager import CLIPModelManager
 from src.analyzers.metric_calculator import MetricCalculator
+from src.constants.genre_weights import GenreWeights
 from src.models.analyzer_config import AnalyzerConfig
-from src.models.genre_weights import GenreWeights
 
 
 @pytest.fixture(autouse=True)
